bootchartd: add getty to EXIT_PROC and save boot_time + exit_proc
A common metric on embedded systems is to measure time from power-on to getty
launches.  This commit adds getty and a few other things in an attempt to get
a better value for Boot Time in the generated graph.

We now make a distinction between the total runtime (duration), which can be
tweaked with the new SETTLE_TIME variable, and boot time, which is the time
from boot until an EXIT_PROC is detected.

When we find the EXIT_PROC, we save a new measurement boot_time, defined as
the time until the EXIT_PROC is found.  We also save the PID and name of the
EXIT_PROC in exit_proc.

While working on this we found that the collector does not respond to signals
when ptraced, so it must detach first.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>